chore: moved conn middleware to conn package
