+++ internal/conn/middleware.go
@@ -1,4 +1,4 @@
-package tools
+package conn
 
 import (
 	"context"
@@ -6,25 +6,24 @@ import (
 	"log"
 	"net/http"
 
-	"github.com/SpatiumPortae/portal/internal/conn"
 	"github.com/gorilla/websocket"
 )
 
 type connKey struct{}
 
-func WithConn(ctx context.Context, conn conn.Conn) context.Context {
+func WithConn(ctx context.Context, conn Conn) context.Context {
 	return context.WithValue(ctx, connKey{}, conn)
 }
 
-func FromContext(ctx context.Context) (conn.Conn, error) {
-	conn, ok := ctx.Value(connKey{}).(conn.Conn)
+func FromContext(ctx context.Context) (Conn, error) {
+	conn, ok := ctx.Value(connKey{}).(Conn)
 	if !ok {
 		return nil, errors.New("unable to get Conn from context")
 	}
 	return conn, nil
 }
 
-func WebsocketMiddleware() func(http.Handler) http.Handler {
+func Middleware() func(http.Handler) http.Handler {
 	wsUpgrader := websocket.Upgrader{}
 
 	return func(next http.Handler) http.Handler {
@@ -34,7 +33,7 @@ func WebsocketMiddleware() func(http.Handler) http.Handler {
 				log.Println("failed to upgrade connection:", err)
 				return
 			}
-			next.ServeHTTP(w, r.WithContext(WithConn(r.Context(), &conn.WS{Conn: wsConn})))
+			next.ServeHTTP(w, r.WithContext(WithConn(r.Context(), &WS{Conn: wsConn})))
 		})
 	}
 }
