+++ internal/rendezvous/routes.go
@@ -1,9 +1,9 @@
 package rendezvous
 
-import "github.com/SpatiumPortae/portal/tools"
+import "github.com/SpatiumPortae/portal/internal/conn"
 
 func (s *Server) routes() {
-	s.router.Use(tools.WebsocketMiddleware())
+	s.router.Use(conn.Middleware())
 	s.router.HandleFunc("/establish-sender", s.handleEstablishSender())
 	s.router.HandleFunc("/establish-receiver", s.handleEstablishReceiver())
 }
